+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## [0.0.7](../../releases/tag/v0.0.7) - Unreleased
 
-- ...
+### Fixes
+
+- selector handling in `enqueue_links` in `BeautifulSoupCrawler`
 
 ## [0.0.6](../../releases/tag/v0.0.6) - 2024-06-25